# docs/swagger.yaml
definitions:
...
  api_response.CreateTodoResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
      todo:
        $ref: '#/definitions/models.Todo'
    type: object
...
  api_response.DeleteTodoResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  api_response.GetTodoResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
      todo:
        $ref: '#/definitions/models.Todo'
    type: object
  api_response.GetTodosResponse:
    properties:
      message:
        type: string
      pagination:
        $ref: '#/definitions/types.Pagination'
      status:
        type: integer
      success:
        type: boolean
      todos:
        items:
          $ref: '#/definitions/models.Todo'
        type: array
    type: object
...
  api_response.UpdateTodoResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
      todo:
        $ref: '#/definitions/models.Todo'
    type: object
...
  models.Todo:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_completed:
        type: boolean
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  models.UpdateTodoRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      is_completed:
        type: boolean
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - title
    type: object
...
  types.Pagination:
    properties:
      page:
        type: integer
      pages:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: API documentation
  title: Example API
  version: "1.0"
paths:
...
  /api/v1/todos:
    get:
      description: Retrieve todos with optional pagination (page, size) and filter
        by is_completed
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: size
        type: integer
      - description: Filter by completion status
        in: query
        name: is_completed
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.GetTodosResponse'
      security:
      - BearerAuth: []
      summary: Get todos with pagination and filters
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Create a new todo item
      parameters:
      - description: Todo object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.CreateTodoResponse'
        "400":
          description: Invalid JSON
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new todo
      tags:
      - todos
  /api/v1/todos/{id}:
    delete:
      description: Delete a specific todo by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.DeleteTodoResponse'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Todo not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete todo by ID
      tags:
      - todos
    get:
      description: Fetch a specific todo by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.GetTodoResponse'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Todo not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get todo by ID
      tags:
      - todos
    patch:
      consumes:
      - application/json
      description: Update a todo item by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Todo object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTodoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_response.UpdateTodoResponse'
        "400":
          description: Invalid JSON
          schema:
            type: string
        "404":
          description: Todo not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update an existing todo
      tags:
      - todos
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"